# docs-site/openapi/ai-engine.openapi.yaml
openapi: 3.0.3
info:
  title: Laravel AI Engine API
  version: 1.0.0
  description: |
    Interactive API reference for the Laravel AI Engine package.

    This spec covers the current v1 application APIs and the node federation
    contract.
servers:
  - url: http://ai.test
    description: Local test host
  - url: http://127.0.0.1:8000
    description: Laravel artisan serve
...
paths:
...
  /api/v1/data-collector/message:
    post:
      tags: [Data Collectors]
      summary: Continue a guided collector session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataCollectorMessageRequest'
      responses:
        '200':
          description: Collector response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataCollectorMessageResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/data-collector/data/{sessionId}:
    get:
      tags: [Data Collectors]
      summary: Get collected session data
      parameters:
        - $ref: '#/components/parameters/SessionIdCamel'
      responses:
        '200':
          description: Session data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session_id:
                    type: string
                  data:
                    $ref: '#/components/schemas/JsonObject'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/autonomous-collector/data/{sessionId}:
    get:
      tags: [Autonomous Collectors]
      summary: Get autonomous collector data
      parameters:
        - $ref: '#/components/parameters/SessionIdCamel'
      responses:
        '200':
          description: Session data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session_id:
                    type: string
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/JsonObject'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    nodeTokenAuth:
      type: apiKey
      in: header
      name: X-Node-Token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    SessionIdCamel:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/JsonObject'
        message:
          type: string
        errors:
          type: object
          additionalProperties: true
        trace:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        errors:
          type: object
          additionalProperties: true
...
    DataCollectorMessageRequest:
      type: object
      required: [session_id, message]
      properties:
        session_id:
          type: string
        message:
          type: string
        config_name:
          type: string
        engine:
          type: string
        model:
          type: string
    DataCollectorMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        actions:
          type: array
          items:
            $ref: '#/components/schemas/JsonObject'
        metadata:
          $ref: '#/components/schemas/JsonObject'
        config_name:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        is_complete:
          type: boolean
        is_cancelled:
          type: boolean
        requires_confirmation:
          type: boolean
        allows_enhancement:
          type: boolean
        current_field:
          type: string
          nullable: true
        collected_fields:
          type: array
          items:
            type: string
        remaining_fields:
          type: array
          items:
            type: string
        fields:
          type: array
          items:
            type: string
        progress:
          type: number
        data:
          $ref: '#/components/schemas/JsonObject'
        summary:
          type: string
          nullable: true
        action_summary:
          type: string
          nullable: true
        result:
          $ref: '#/components/schemas/JsonObject'
        generated_output:
          $ref: '#/components/schemas/JsonObject'
        config:
          $ref: '#/components/schemas/JsonObject'
...